requirements.txt
Flask==3.0.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
SQLAlchemy>=2.0.40
Werkzeug==3.0.1
APScheduler==3.10.4
email-validator==2.1.0
boto3==1.34.0
psycopg2-binary==2.9.9

